+++ service/controller/resource/organization/create.go
@@ -5,9 +5,14 @@ import (
 	"fmt"
 	"strings"
 
+	"github.com/giantswarm/apiextensions/v3/pkg/apis/security/v1alpha1"
+	"github.com/giantswarm/apiextensions/v3/pkg/label"
 	companyclient "github.com/giantswarm/companyd-client-go"
 	"github.com/giantswarm/microerror"
+	corev1 "k8s.io/api/core/v1"
 	apierrors "k8s.io/apimachinery/pkg/api/errors"
+	"k8s.io/apimachinery/pkg/types"
+	ctrl "sigs.k8s.io/controller-runtime/pkg/client"
 
 	"github.com/giantswarm/organization-operator/service/controller/key"
 )
@@ -25,6 +30,11 @@ func (r *Resource) EnsureCreated(ctx context.Context, obj interface{}) error {
 		}
 	}
 
+	err = r.ensureOrganizationHasSubscriptionIdAnnotation(ctx, org)
+	if err != nil {
+		return microerror.Mask(err)
+	}
+
 	orgNamespace := newOrganizationNamespace(org.Name)
 	r.logger.LogCtx(ctx, "level", "debug", "message", fmt.Sprintf("creating organization namespace %#q", orgNamespace.Name))
 
@@ -37,8 +47,8 @@ func (r *Resource) EnsureCreated(ctx context.Context, obj interface{}) error {
 
 	r.logger.LogCtx(ctx, "level", "debug", "message", fmt.Sprintf("created organization namespace %#q", orgNamespace.Name))
 
-	org.Status.Namespace = orgNamespace.Name
-	err = r.k8sClient.CtrlClient().Status().Update(ctx, &org)
+	patch := []byte(fmt.Sprintf(`{"status":{"namespace": "%s"}}`, orgNamespace.Name))
+	err = r.k8sClient.CtrlClient().Patch(ctx, &org, ctrl.RawPatch(types.MergePatchType, patch))
 	if err != nil {
 		return microerror.Mask(err)
 	}
@@ -60,4 +70,56 @@ func (r *Resource) EnsureCreated(ctx context.Context, obj interface{}) error {
 	r.logger.LogCtx(ctx, "level", "debug", "message", fmt.Sprintf("created legacy organization %#q", legacyOrgName))
 
 	return nil
+}
+
+func (r *Resource) ensureOrganizationHasSubscriptionIdAnnotation(ctx context.Context, organization v1alpha1.Organization) error {
+	r.logger.LogCtx(ctx, "level", "debug", "message", fmt.Sprintf("ensuring organization %q has subscriptionid annotation", organization.Name))
+	// Retrieve secret related to this organization.
+	secret, err := findSecret(ctx, r.k8sClient.CtrlClient(), organization.Name)
+	if IsSecretNotFound(err) {
+		// We don't want this error to block execution so we still return nil and just log the problem.
+		r.logger.LogCtx(ctx, "level", "warning", "message", fmt.Sprintf("unable to find a secret for organization %s. Cannot set subscriptionid annotation", organization.Name))
+		return nil
+	} else if err != nil {
+		return microerror.Mask(err)
+	}
+
+	// The subscription id field is missing in non azure installations so it's ok.
+	if subscription, ok := secret.Data["azure.azureoperator.subscriptionid"]; ok && len(subscription) > 0 {
+		r.logger.LogCtx(ctx, "level", "debug", "message", fmt.Sprintf("setting subscriptionid annotation to %q for organization %q", string(subscription), organization.Name))
+		patch := []byte(fmt.Sprintf(`{"metadata":{"annotations":{"subscription": "%s"}}}`, string(subscription)))
+		err = r.k8sClient.CtrlClient().Patch(ctx, &organization, ctrl.RawPatch(types.MergePatchType, patch))
+		if err != nil {
+			return microerror.Mask(err)
+		}
+	} else {
+		r.logger.LogCtx(ctx, "level", "debug", "message", fmt.Sprintf("azure.azureoperator.subscriptionid field not found or empty in secret %q", secret.Name))
+	}
+
+	return nil
+}
+
+func findSecret(ctx context.Context, client ctrl.Client, orgName string) (*corev1.Secret, error) {
+	// Look for a secret with labels "app: credentiald" and "giantswarm.io/organization: org"
+	secrets := &corev1.SecretList{}
+
+	err := client.List(ctx, secrets, ctrl.MatchingLabels{"app": "credentiald", label.Organization: orgName})
+	if err != nil {
+		return nil, microerror.Mask(err)
+	}
+
+	if len(secrets.Items) > 0 {
+		return &secrets.Items[0], nil
+	}
+	secret := &corev1.Secret{}
+
+	// Organization-specific secret not found, use secret named "credential-default".
+	err = client.Get(ctx, ctrl.ObjectKey{Namespace: "giantswarm", Name: "credential-default"}, secret)
+	if apierrors.IsNotFound(err) {
+		return nil, microerror.Maskf(secretNotFoundError, "Unable to find secret for organization %s", orgName)
+	} else if err != nil {
+		return nil, microerror.Mask(err)
+	}
+
+	return secret, nil
 }