+++ service/controller/resource/organization/error.go
@@ -4,6 +4,15 @@ import (
 	"github.com/giantswarm/microerror"
 )
 
+var secretNotFoundError = &microerror.Error{
+	Kind: "secretNotFoundError",
+}
+
+// IsSecretNotFound asserts secretNotFoundError.
+func IsSecretNotFound(err error) bool {
+	return microerror.Cause(err) == secretNotFoundError
+}
+
 var invalidConfigError = &microerror.Error{
 	Kind: "invalidConfigError",
 }